+++ tests/testthat/test-cc.R
@@ -16,17 +16,20 @@ test_that("cc with atomic vectors", {
 })
 
 test_that("cc errors", {
-  expect_error(cc(1, conj = 1), "^`conj` must be a character scalar[.]$")
-  expect_error(cc(1, conj = character(0)), "^`conj` must be a character scalar[.]$")
-  expect_error(cc(1, conj = as.character(1:2)), "^`conj` must be a character scalar[.]$")
+  expect_error(cc(1, conj = 1), 
+               "^`conj` must be a string [(]non-missing character scalar[)][.]$")
+  expect_error(cc(1, conj = character(0)), 
+               "^`conj` must be a string [(]non-missing character scalar[)][.]$")
+  expect_error(cc(1, conj = as.character(1:2)), 
+               "^`conj` must be a string [(]non-missing character scalar[)][.]$")
   expect_error(cc(1, brac = 1), 
-               "^`brac` must be a character vector of length 1 or 2[.]$")
+               "^`brac` must inherit from class 'character'[.]$")
   expect_error(cc(1, brac = character(0)), 
-               "^`brac` must be a character vector of length 1 or 2[.]$")
+               "^`length[(]brac[)]` must be between 1 and 2, not 0[.]$")
   expect_error(cc(1, brac = as.character(1:3)), 
-               "^`brac` must be a character vector of length 1 or 2[.]$")
+               "^`length[(]brac[)]` must be between 1 and 2, not 3[.]$")
   expect_error(cc(1, sep = as.character(1:3)), 
-               "^`sep` must be a character scalar[.]$")
+               "^`sep` must be a string [(]non-missing character scalar[)][.]$")
 })
 
 test_that("cc with character", {
@@ -56,4 +59,13 @@ test_that("cc with random objects", {
 
 test_that("cc with sep", {
   expect_identical(cc(1:3, sep = "|"), "1|2, 3")
+})
+
+test_that("cc with ellipsis", {
+  expect_identical(cc(1:10), c("1, 2, 3, 4, 5, 6, 7, 8, 9, 10"))
+  expect_identical(cc(1:10, ellipsis = 3), c("1, ..., 10"))
+  expect_error(cc(1:10, ellipsis = 2), 
+               "`ellipsis` must be greater than or equal to 3, not 2[.]")
+  expect_error(cc(1:10, ellipsis = 2.5), 
+               "`ellipsis` must be a whole number [(]non-missing integer scalar or double equivalent[)][.]")
 })